+++ packages/client/src/features/accessibility/search/search-palette-module.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019-2023 EclipseSource and others.
+ * Copyright (c) 2023 Business Informatics Group (TU Wien) and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -13,7 +13,18 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-export * from './direct-task-editing/direct-task-editor';
-export * from './model';
-export * from './workflow-diagram-module';
-export * from './workflow-views';
+
+import { ContainerModule } from 'inversify';
+import { bindAsService, BindingContext, TYPES } from '~glsp-sprotty';
+import { SearchAutocompletePalette } from './search-palette';
+import { SearchAutocompletePaletteTool } from './search-tool';
+
+export const searchPaletteModule = new ContainerModule((bind, _unbind, isBound, rebind) => {
+    const context = { bind, isBound, rebind };
+    configureSearchPaletteModule(context);
+});
+
+export function configureSearchPaletteModule(context: Pick<BindingContext, 'bind'>): void {
+    bindAsService(context, TYPES.IUIExtension, SearchAutocompletePalette);
+    bindAsService(context, TYPES.IDefaultTool, SearchAutocompletePaletteTool);
+}